+++ yamlgraph/utils/__init__.py
@@ -7,14 +7,6 @@ from yamlgraph.utils.expressions import (
     resolve_template,
 )
 from yamlgraph.utils.json_extract import extract_json
-from yamlgraph.utils.langsmith import (
-    get_client,
-    get_latest_run_id,
-    get_project_name,
-    get_run_url,
-    is_tracing_enabled,
-    print_run_tree,
-)
 from yamlgraph.utils.logging import get_logger, setup_logging
 from yamlgraph.utils.prompts import load_prompt, load_prompt_path, resolve_prompt_path
 from yamlgraph.utils.template import extract_variables, validate_variables
@@ -28,13 +20,6 @@ __all__ = [
     "resolve_template",
     # JSON extraction
     "extract_json",
-    # LangSmith
-    "get_client",
-    "get_project_name",
-    "is_tracing_enabled",
-    "get_latest_run_id",
-    "print_run_tree",
-    "get_run_url",
     # Logging
     "get_logger",
     "setup_logging",